@@ -7,17 +7,18 @@
   "devDependencies": {
     "@commitlint/cli": "17",
     "@commitlint/config-conventional": "17",
-    "@midwayjs/cli": "^2.1.1",
     "@midwayjs/mock": "^3.12.2",
+    "@midwayjs/swagger": "^3.12.2",
     "@midwayjs/web": "^3.12.2",
     "@rollup/plugin-commonjs": "25",
     "@rollup/plugin-node-resolve": "15",
     "@types/mocha": "10",
     "@types/node": "20",
-    "@waiting/eslint-config": "^8.8.0",
+    "@waiting/eslint-config": "^9.0.0",
     "@waiting/shared-core": "^21.5.0",
     "@waiting/shared-types": "^21.5.0",
     "autocannon": "7",
+    "c8": "^8.0.1",
     "coveralls": "3",
     "cross-env": "7",
     "debug": "^4.3.4",
@@ -25,6 +26,7 @@
     "madge": "^6.1.0",
     "mocha": "10",
     "mocha-lcov-reporter": "1",
+    "mwtsc": "^1.0.1",
     "nx": "16",
     "rollup-plugin-dts": "6",
     "ts-node": "^10.9.1",
@@ -35,7 +37,7 @@
     "zx": "7"
   },
   "engines": {
-    "node": ">=16.13.0"
+    "node": ">=18.12.0"
   },
   "workspaces": [
     "packages/*"
@@ -49,8 +51,6 @@
     "build": "sh .scripts/build.sh",
     "ci": "npm run cov",
     "clean": "lerna clean --loglevel error --yes && lerna run clean",
-    "clean:cache": "rm -f packages/*/.eslintcache packages/*/.tsbuildinfo",
-    "clean:dist": "rm -rf packages/*/dist/ packages/*/.tsbuildinfo",
     "cov": "sh .scripts/cov.sh",
     "lint": "sh .scripts/lint.sh",
     "lint:cycle": "lerna exec -- madge --circular --extensions ts src",
@@ -63,7 +63,7 @@
     "pub:canary": "bash .scripts/publish.sh --canary",
     "pub:beta": "bash .scripts/publish.sh --dist-tag beta --force-publish=*",
     "pub:next": "bash .scripts/publish.sh --dist-tag next",
-    "purge": "npm run clean && lerna run purge && nx reset && rm node_modules -rf && rm package-lock.json -f && npm run refresh",
+    "purge": "npm run clean && nx reset && rm node_modules -rf && rm package-lock.json -f && npm run refresh",
     "repo:init": "sh .githooks/init-repo.sh",
     "refresh": "npm i && npm run build",
     "test": "lerna run test",
