@@ -1,5 +1,5 @@
 /**
- * 搜索指定目录以 file.example 文件为基础生成不带后缀的文件为不带 .example 后缀的文件
+ * 搜索指定目录以 file.example 文件为基础生成 去除结尾 .example
  */
 import { join } from 'node:path'
 import { cp, readdir, stat } from 'node:fs/promises'
