@@ -5,13 +5,13 @@ import {
 } from '@midwayjs/core'
 import { Trace } from '@mwcp/otel'
 
-import { Config, ConfigKey, Msg } from '../lib/types'
+import { Config, ConfigKey, Msg } from '##/lib/types.js'
 
 
 @Controller(`/${ConfigKey.namespace}/hello`)
 export class DefaultComponentController {
 
-  @_Config(ConfigKey.config) readonly config: Config
+  @_Config(ConfigKey.clientConfig) readonly config: Config
 
   @Trace()
   @Get('/')
