@@ -1,3 +1,3 @@
 
-export * from './default.controller'
+export * from './default.controller.js'
 
