@@ -1,6 +1,6 @@
 
-export * from './config'
-export * from './config.types'
-export * from './db.model'
-export * from './types'
+export * from './config.js'
+export * from './config.types.js'
+export * from './db.model.js'
+export * from './types.js'
 
