@@ -1,5 +1,5 @@
 
-export * from './task-queue.repo'
-export * from './task-result.repo'
-export * from './task-log.repo'
+export * from './task-queue.repo.js'
+export * from './task-result.repo.js'
+export * from './task-log.repo.js'
 
