@@ -1,4 +1,4 @@
 // 导出的是常量，而非类型
 
-export * from './tm.schema'
+export * from './tm.schema.js'
 
