@@ -1,26 +1,45 @@
 import { IncomingHttpHeaders } from 'node:http'
+import { join } from 'node:path'
 
-import type { Application, IMidwayContainer } from '@mwcp/share'
+import {
+  Application,
+  IMidwayContainer,
+  JsonResp,
+} from '@mwcp/share'
+import { genCurrentDirname } from '@waiting/shared-core'
 import supertest, { SuperTest } from 'supertest'
 
-import { ClientService } from '~/index'
+import { ClientService } from '##/index.js'
 import {
   TaskLogRepository,
   TaskQueueRepository,
   TaskResultRepository,
-} from '~/repo/index.repo'
-import { TaskAgentService, TaskQueueService } from '~/service/index.service'
+} from '##/repo/index.repo.js'
+import { TaskAgentService, TaskQueueService } from '##/service/index.service.js'
 
 
-const CI = !! process.env['CI']
+export const testDir = genCurrentDirname(import.meta.url)
+export const baseDir = join(testDir, '..')
+
+const CI = !! (process.env['CI']
+  || process.env['MIDWAY_SERVER_ENV'] === 'unittest'
+  || process.env['MIDWAY_SERVER_ENV'] === 'local'
+  || process.env['NODE_ENV'] === 'unittest'
+  || process.env['NODE_ENV'] === 'local'
+)
+
 export type TestResponse = supertest.Response
-export interface TestRespBody {
+export type TestRespBody = JsonResp<RespData>
+export interface RespData {
   header: IncomingHttpHeaders
   url: string
   cookies: unknown
 }
 
 export interface TestConfig {
+  baseDir: string
+  testDir: string
+  testAppDir: string
   CI: boolean
   app: Application
   container: IMidwayContainer
@@ -33,8 +52,14 @@ export interface TestConfig {
   retRepo: TaskResultRepository
   tm: ClientService
 }
+
+const testAppDir = join(testDir, 'fixtures', 'base-app')
 export const testConfig = {
+  baseDir,
+  testDir,
+  testAppDir,
   CI,
   host: '',
+  httpRequest: {},
 } as TestConfig
 
